Feat : 마이페이지 뱃지 API MongoDB 연동 진행중

+++ backend/src/main/java/com/ssafy/cnnect/badge/dto/BadgeListResponseDto.java
@@ -12,6 +12,6 @@ public class BadgeListResponseDto {
     private String badgeName;
     private String badgeCondition;
     private String badgeCategory;
-    private int badgeCount;
-    private int badgeScore;
+    private Long badgeCount;
+    private Long badgeScore;
 }
